+++ backend/src/looksyk/builtinpage/mod.rs
@@ -1,6 +1,7 @@
 pub mod asset_metainfo_table;
 pub mod assets_overview;
 mod generating_page_util;
+pub mod help_page;
 pub mod journal_overview;
 pub mod page_not_found;
 pub mod user_page_overview;