+++ backend/src/looksyk/queries/mod.rs
@@ -5,5 +5,6 @@ pub mod insert_file_content;
 pub mod pagehierarchy;
 pub mod references_to;
 pub mod todo;
+pub mod todo_progress;
 
 mod basic;